# Conflicts:
#	features/common/common_tradein/build.gradle
#	features/communication/broadcast_message/build.gradle
#	features/digital/digital_browse/build.gradle
#	features/discovery/productcard_options/build.gradle
#	features/fcm_common/build.gradle
#	features/merchant/product_detail/build.gradle
#	features/promo/tokopoints/build.gradle
#	features/transaction/sellerorder/build.gradle
#	features/user/notifcenter/build.gradle
#	features/user/settingbank/build.gradle
#	features/user/user_identification/build.gradle
#	features/user/user_identification_common/build.gradle
#	libraries/analytics/build.gradle
#	tkpdabstraction/build.gradle
#	tkpdabstraction/src/main/res/values/colors.xml
#	tkpdabstraction/src/main/res/values/dimens.xml

+++ features/payment/withdraw/src/main/java/com/tokopedia/withdraw/view/listener/WithdrawPasswordContract.java
@@ -29,6 +29,6 @@ public class WithdrawPasswordContract {
 
     public interface Presenter extends CustomerPresenter<View> {
 
-        void doWithdraw(int withdrawal, BankAccount bankAccount, String password, boolean isSellerWithdrawal, String programName);
+        void doWithdraw(long withdrawal, BankAccount bankAccount, String password, boolean isSellerWithdrawal, String programName);
     }
 }
